requirements.txt
undetected-chromedriver
termcolor
py_imessage
selenium
Pillow
pytesseract
requests
